requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
mysql-connector-python==9.0.0